Moved RouteDisplay.vue and Routes.vue to the dashboard_items directory where Dashboard.vue resides and accesses them Co-authored-by: Kerolos Nesem <nesemkerolos@gmail.com> Co-authored-by: Megan Nadkarni <megan.nadkarni@gmail.com> Co-authored-by: Julia Bakerink < juliabakerink@gmail.com>
